+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to the
 - Add up to the multiplication examples in Chapter 5
 - Add up to the pretty-print example in Chapter 5
 - Add up to the calculator example in Chapter 6
+- Add up to the teletype and filesystem example in Chapter 7
 
 ## 0.1.0.0 - YYYY-MM-DD